Refactor package structure and update tests
Renamed License3jSpringApplication.java to a new package structure and removed unnecessary ComponentScan annotation. Updated NewLicenseTest to use @ SpringBootTest and @ AutoConfigureMockMvc, replacing @ SpringJUnitWebConfig and @ MockitoSpyBean
