+++ src/main/java/org/egg/license3j/api/License3jSpringApplication.java
@@ -1,11 +1,9 @@
-package org.egg.license3j.api.main;
+package org.egg.license3j.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@ComponentScan(basePackages = "org.egg.license3j.api")
 public class License3jSpringApplication {
 
 	public static void main(String[] args) {
